move ChecksumHelper
