+++ build-extensions/src/main/kotlin/eu/cloudnetservice/cloudnet/gradle/plugins/updater/GenerateUpdaterInformationTask.kt
@@ -20,7 +20,7 @@ import com.google.gson.Gson
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
-import eu.cloudnetservice.cloudnet.gradle.ChecksumHelper
+import eu.cloudnetservice.cloudnet.gradle.util.ChecksumHelper
 import eu.cloudnetservice.cloudnet.gradle.Versions
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.ConfigurableFileCollection
