+++ build-extensions/src/main/kotlin/eu/cloudnetservice/cloudnet/gradle/tasks.kt
@@ -18,6 +18,7 @@
 
 package eu.cloudnetservice.cloudnet.gradle
 
+import eu.cloudnetservice.cloudnet.gradle.util.ChecksumHelper
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
